refactor: with-connect example
