--- examples/with-cloud/package.json
@@ -1,44 +1,0 @@
-{
-    "name": "with-cloud",
-    "version": "3.20.1",
-    "private": true,
-    "dependencies": {
-        "@refinedev/antd": "^5.3.12",
-        "@refinedev/core": "^4.5.8",
-        "@refinedev/cli": "^2.1.2",
-        "@refinedev/react-router-v6": "^4.1.0",
-        "@refinedev/simple-rest": "^4.1.0",
-        "@refinedev/cloud": "^9.1.0",
-        "@refinedev/antd-audit-log": "^4.1.0",
-        "@types/node": "^12.20.11",
-        "@types/react": "^18.0.0",
-        "@types/react-dom": "^18.0.0",
-        "react": "^18.0.0",
-        "react-dom": "^18.0.0",
-        "@uiw/react-md-editor": "^3.19.5",
-        "react-scripts": "^5.0.0",
-        "cross-env": "^7.0.3",
-        "typescript": "^4.7.4",
-        "web-vitals": "^1.1.1",
-        "antd": "^5.0.5",
-        "@ant-design/icons": "^5.0.1"
-    },
-    "scripts": {
-        "start": "cross-env DISABLE_ESLINT_PLUGIN=true react-scripts start",
-        "build": "cross-env DISABLE_ESLINT_PLUGIN=true react-scripts build",
-        "eject": "react-scripts eject",
-        "refine": "refine"
-    },
-    "browserslist": {
-        "production": [
-            ">0.2%",
-            "not dead",
-            "not op_mini all"
-        ],
-        "development": [
-            "last 1 chrome version",
-            "last 1 firefox version",
-            "last 1 safari version"
-        ]
-    }
-}
