+++ examples/with-connect/.gitignore
