+++ examples/with-connect/public/favicon.ico
