+++ examples/with-connect/public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "refine cloud example",
-  "name": "refine cloud example",
+  "short_name": "refine connect example",
+  "name": "refine connect example",
   "icons": [
     {
       "src": "favicon.ico",
