+++ examples/with-connect/src/index.tsx
