+++ examples/with-connect/src/interfaces/index.d.ts
