+++ examples/with-connect/src/pages/categories/index.tsx
