+++ examples/with-connect/src/pages/posts/index.tsx
