+++ examples/with-connect/src/pages/posts/list.tsx
