+++ examples/with-connect/src/pages/posts/show.tsx
