+++ examples/with-connect/src/pages/products/index.tsx
