+++ examples/with-connect/src/utility/normalize.ts
