+++ examples/with-connect/tsconfig.json
